requirements.txt
folium==0.15.1
matplotlib==3.3.4
mapclassify==2.5.0
plotly==5.18.0
urllib3==1.26.5
geopandas
pandas
numpy
streamlit
streamlit_folium
streamlit_option_menu
streamlit_authenticator==0.2.3
streamlit_echarts
streamlit_aggrid
streamlit_dynamic_filters
altair==5.0.1
PyYAML==6
xlsxwriter==3.0.8
openpyxl==3.0.10
locales
streamlit-agraph==0.0.45


